requirements.txt
# Stock Portfolio Management Application - Python Dependencies

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Financial data
yfinance>=0.2.28

# Visualization
matplotlib>=3.7.0

# HTTP requests for currency conversion
requests>=2.31.0

# HTML parsing for web scraping
beautifulsoup4>=4.12.0

# ODF (OpenDocument) file parsing for .ods files
odfpy>=1.4.1

# AI Agent dependencies
anthropic>=0.21.0      # Anthropic Claude API
openai>=1.12.0         # OpenAI GPT API (optional)
google-generativeai>=0.3.0  # Google Gemini API (optional)
pypdf>=4.0.0           # PDF parsing for report analysis

# Remote data fetching (optional - only needed for specific protocols)
# paramiko>=2.11.0  # Uncomment for SSH/SFTP support
# boto3>=1.26.0     # Uncomment for S3 support

# Note: curses is included in the Python standard library on Unix/Linux systems
# For Windows users, install windows-curses:
# windows-curses>=2.3.0

# System packages (install via package manager):
# Ubuntu/Debian: sudo apt-get install python3-tk
# Fedora/RHEL:   sudo dnf install python3-tkinter
# macOS:         Included with Python installation
# (Required for matplotlib TkAgg backend to display graphs)
